Started implementation for Hero and Medic, added makefile and macros.h
Started the implementation of the Hero and Medic classes, although some of
the functions are still only stubs. Especially the "cure" function needs
deeper thought, as checking the disease ID for cards would need to be done
indirectly, first looking at the city and then the city disease id.

The other hero specialisms need to be implemented.

Added a sample makefile, which needs to be adjusted to be usable, as there
currently is no main function in the project -- in the makefile, I have
assumed that it will be contained in a file "main.cpp", but this file does
not exist yet. However, apart from potential additions or changes needed to
the first line specifying object variables ("OBJ= ..."), the makefile should
work as is once the main.cpp (or otherwise) is implemented.

macros.h includes preprocessor directives (macros) that should be useful
in the project. Currently it has only a limited number of these, relating
to disease IDs and disease status, but the idea is to add more.

 Changes to be committed:
	new file:   Medic.cpp
	modified:   Medic.h
	modified:   city.h
	new file:   hero.cpp
	modified:   hero.h
	new file:   macros.h
	new file:   makefile
	modified:   pcard.h
	modified:   world.h