GUI layout mockup - trying out button and progress bar placement (WIP)
Early tests with the layout of the possible GUI. The thing is still
clunky, and none of the buttons on the board connect to any functions
in the underlying game system. However, this gives a basis for an
early version of the GUI, so we can start testing things and improve.