Changed Window to HandWindow, added Specialwindow class for special action
handwindow.cpp now defines a window only for the players hand (and in
the future the hands of other players), while separate classes are
defined for other windows.

Added a specialwindow to provide a window for controlling special actions,
for the classes that have such actions. This is still an early prototype.

 Changes to be committed:
	modified:   Contagion.pro
	renamed:    window.cpp -> handwindow.cpp
	new file:   handwindow.h
	modified:   main.cpp
	modified:   mainwindow.cpp
	modified:   mainwindow.h
	new file:   specialwindow.cpp
	new file:   specialwindow.h
	deleted:    window.h