Implementing city buttons (for movement) as radio buttons
Added a small set of radio buttons to enable movement choices
to cities. The plan is that the network of cities is represented
in the map on the backrgound, and the buttons appear on top of the
map (over the cities) when a move is to be made (so appear as an
overlay window). Currently the functionality is being tested in
the mainwindow, but this should be moved into a "movewindow" or
something similar (which will control the overlay).

The stylesheet "style.qss" is used to control the style of the
widgets (buttons, menus etc.), like the size of the radio buttons,
the colour of buttons and menus etc.

Also added small fixes to specialwindow.cpp

 Changes to be committed:
	modified:   Contagion.pro
	modified:   mainwindow.cpp
	new file:   resources/images/radiobutton_checked.png
	new file:   resources/images/radiobutton_checked_hover.png
	new file:   resources/images/radiobutton_checked_pressed.png
	new file:   resources/images/radiobutton_unchecked.png
	new file:   resources/images/radiobutton_unchecked_hover.png
	new file:   resources/images/radiobutton_unchecked_pressed.png
	new file:   resources/style.qss
	modified:   specialwindow.cpp
	modified:   specialwindow.h