Added the move window overlay "movewindow"
The move window overlay "movewindow" now includes the radio buttons
for cities, for the purpose of movement.

The movewindow class constructor generates the cities based
on the cities.dat file, although the coordinates don't match the map
properly as the map projection differs from that used by the map
which the coordinates are based on. It is possible that a separate data
file should be created specifically for this map, unless a good conversion
is found.

 Changes to be committed:
	modified:   Contagion.pro
	modified:   mainwindow.cpp
	modified:   mainwindow.h
	new file:   movewindow.cpp
	new file:   movewindow.h