Created a class for confirming actions; added functionality to movewindow
Added "confirmwindow" class for creating an overlay window that
asks to confirm player choices, e.g. movement to city etc.
Currently this is enabled only for the basic "move" selection, and
not for e.g. special actions or flights.

Also added to the functionality of movewindow, so now pressing the
city buttons creates signals that can then be handled to make the
moves happen. The implemented functionality is still only between the
GUI elements, although tying in the game logic functions does not
require much more work.

 Changes to be committed:
	modified:   Contagion.pro
	new file:   confirmwindow.cpp
	new file:   confirmwindow.h
	modified:   handwindow.cpp
	modified:   main.cpp
	modified:   mainwindow.cpp
	modified:   mainwindow.h
	modified:   movewindow.cpp
	modified:   movewindow.h