Implemented the basic actions from the Move overlay
The move button, and city buttons in the overlay, now interact with
the World, City and Hero classes to actually result in a player move.

There are still some fancier things to be added, like making the city
buttons only clickable for the cities that are next to the current
location of the hero, but that's less relevant at the moment, as
many other features are completely missing.

Also implemented a counter displaying the number of actions a player
has remaining. It is simple to use, but the refresh has to be coded
for each action separately, as it doesn't actively track actions.
It could be made to react to changes in number of actions directly,
if Hero forced an update every time the number of "moves" changes.
However, this would mean breaking the separation between the GUI
and the Hero class, which we do not want.

 Changes to be committed:
	modified:   Contagion.pro
	modified:   resources/style.qss
	modified:   src/gui/main.cpp
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h
	new file:   src/gui/movebuttons.cpp
	new file:   src/gui/movebuttons.h
	modified:   src/gui/movewindow.cpp
	modified:   src/gui/movewindow.h
	modified:   world.cpp
	modified:   world.h