Added: action counter; end-of-turn window; meeple sprites (simple)
Action counter class (actioncounter) tracks the number of actions
the player has taken since their turn started. The number of actions
remaining is displayed at the top of the screen. When the counter
hits 0, an end-of-turn window is displayed.

End of turn window class (endturnwindow) is an overlay which informs
the player that they have taken all of their actions, and prompts
them to get rid of excess cards in their hand. Currently this is
just a placeholder, with most of the functionalities still to be
implemented.

Meeple sprite class (meeplesprite) is a class for the meeples,
which are implemented as QGraphicsObjects. Currently these are
simple coloured circles, but can be replaced with proper meeples
with animations. Also added QGraphicsScene and QGraphicsView
members to the main window, on which the sprites can be displayed.

 Changes to be committed:
	modified:   Contagion.pro
	modified:   hero.h
	modified:   resources/style.qss
	new file:   src/gui/actioncounter.cpp
	new file:   src/gui/actioncounter.h
	new file:   src/gui/endturnwindow.cpp
	new file:   src/gui/endturnwindow.h
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h
	new file:   src/gui/meeplesprite.cpp
	new file:   src/gui/meeplesprite.h
	modified:   src/gui/movewindow.cpp