Added: disease cube generation to the graphics scene; disinfect button
Added a disinfect button, as this was missing. It's still missing
some of the implementation, including a selection screen for selecting
which colour cubes to remove. These are easy to add, though.

Disease cubes are now drawn onto the graphics scene. This is currently
only done at the start of the game, but should be easy to use the
function to re-draw the cubes when a change (infect or treat) takes
place.
NOTE: There was a bug in the city.cpp, as the constructor used in
world setup was not set to initialize the disease counts (disease_counters)
to zero. This lead to garbage values in cities, and system crash as the
gui disease cube setup function trying to draw billions of cubes.

 Changes to be committed:
	modified:   city.cpp
	modified:   src/gui/endturnwindow.cpp
	modified:   src/gui/endturnwindow.h
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h
	modified:   src/gui/meeplesprite.cpp
	modified:   src/gui/meeplesprite.h
	modified:   src/gui/movewindow.cpp
	modified:   src/gui/movewindow.h
	modified:   src/gui/specialwindow.h