Implemented "Disinfect" button's action; re-did graphical disease cubes
The action of the Disinfect button now works, at least the basic case
with only one type of disease in a city. Will have to test how the
selection of diseases works when multiple diseases are in a city,
which can only happen after an outbreak.

The graphical disease cubes had to be re-done, as it was hard to
control the destruction of the ones implemented using QRectF. For
this, I introduced a new class, "diseasecube". This was required
for re-drawing disease cubes when "disinfect" was done,
as old graphics items (cubes) had to be removed from the scene, and
this was hard as QRectF provides little control for removal (except
via making a list or a vector identifying pointers to them, which
is unnecessarily clumsy). Instead, the new "diseasecube" class
derives from QGraphicsItem, and provides more control -- I used
the "setData() and data()" member functions to store a string that
differentiates cubes from meeples, so meeples are not accidentally
removed from the scene.

 Changes to be committed:
	modified:   Contagion.pro
	modified:   src/gui/confirmwindow.cpp
	new file:   src/gui/diseasecube.cpp
	new file:   src/gui/diseasecube.h
	new file:   src/gui/disinfectwindow.cpp
	new file:   src/gui/disinfectwindow.h
	modified:   src/gui/endturnwindow.cpp
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h
	modified:   src/gui/meeplesprite.h
	modified:   src/gui/movewindow.cpp