Fixed a bug in movewindow.cpp that was caused by changed resource handling
The Move window overlay (movewindow) failed to work with the new
way of handling resources with a .qrc file approach, as a function
was using ifstreams to read data, and the file name system in .qrc
was not compatible with this.

I replaced the problematic function with a better implementation, so
there is no file reads anymore, but it uses the data in the city
objects directly, in world::cities.

 Changes to be committed:
	modified:   src/gui/diseasecube.cpp
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/movewindow.cpp
	modified:   src/gui/movewindow.h