Added: hand-limit check and pop-up message; fixing card window buttons
Added a check for the hand-limit, and a pop-up message box which tells
the player to discard cards if hand-limit exceeded. Currently the hand
limit is tested after each action, although a fuller implementation
would involve a check after every card addition (via drawing or trading).
This should be implemented, but the current version includes the action
that should take place (pop-up message and hand window opening) when
the hand limit is exceeded.
Perhaps the discard window should be implemented separately from the
hand window, show it could created from different windows, like
endturnwindow.

Also fixed the card window buttons, so give only shows for city cards,
and under the conditions when giving a card is permitted.

 Changes to be committed:
	modified:   src/gui/actioncounter.cpp
	modified:   src/gui/actioncounter.h
	modified:   src/gui/cardwindow.cpp
	modified:   src/gui/cardwindow.h
	modified:   src/gui/endturnwindow.cpp
	modified:   src/gui/handwindow.cpp
	modified:   src/gui/handwindow.h
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h
	modified:   src/gui/movewindow.cpp