Added updating to the outbreaks and infection rate progress bars
The outbreaks and infection rate progress bars now update after each
turn end (if there is a need for update).

The infection rate may be better illustrated by a number instead,
like moves remaining (although for the computer game the specific
rate number is not that important, as long as the consequences are
clear).

 Changes to be committed:
	modified:   src/gui/endturnwindow.cpp
	modified:   src/gui/endturnwindow.h
	modified:   src/gui/mainwindow.cpp