Added a player turn indicator on the GUI main window
The player turn is now indicated by a picture/label at the top
of the main window, making it clear which hero is in question,
and which colour the hero is.

The label uses pictures that are changed at the start of each
turn. The current pictures are very basic (consisting of the
hero name on a block of the hero's colour), but can be easily
updated to a fancier look later.

 Changes to be committed:
	new file:   images/cont_planner_turn_bevel.png
	new file:   images/dispatcher_turn_bevel.png
	new file:   images/medic_turn_bevel.png
	new file:   images/op_exp_turn_bevel.png
	new file:   images/quar_spec_turn_bevel.png
	new file:   images/researcher_turn_bevel.png
	new file:   images/scientist_turn_bevel.png
	modified:   resources.qrc
	modified:   src/gui/actioncounter.cpp
	modified:   src/gui/endturnwindow.cpp
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h