Improvements to the background map; changes to special action window
Adding a network of cities to the map on the background (wip - further
improvements needed).
Improvements to the special action button functionality - now it opens
a window with buttons customized based on which hero is played.
The actual functionality will need to be added still - this will likely
involve creating additional selection windows (1. Which meeple to move,
and where? 2. Which city to move to, and which card to discard? etc.).

 Changes to be committed:
	modified:   images/World_map_concept_coloured.xcf
	modified:   src/gui/handwindow.cpp
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/specialwindow.cpp
	modified:   src/gui/specialwindow.h