Added a "flywindow" class, and implemented main flight functionality
The flywindow class derives from movewindow, implementing a different
functionality for displaying city buttons (only shows the cities where
the player can fly to).

The current implementation provides the basic flight functionality.
However, it is currently lacking a way in which the flight type
can be selected if alternatives exist. This could be implemented
either before showing the destination buttons (updating the buttons
based on which type is selected), or when a destination has been
selected, allowing to decide which flight type to choose if
alternatives exist.

 Changes to be committed:
	modified:   Contagion.pro
	new file:   src/gui/flywindow.cpp
	new file:   src/gui/flywindow.h
	modified:   src/gui/handwindow.cpp
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h
	modified:   src/gui/movewindow.cpp
	modified:   src/gui/movewindow.h