Added "flightselector" class, allowing for selection of flight type
The flightselector class provides better functionality to flights,
by allowing the selection of flight type before the "fly window"
is displayed. The available destinations are then displayed based
on the flight type selected. This was judged to be a clear approach
to providing flight type selection when multiple types are available
for a given destination.

The implementation is ready as far as the "flightselector" class is
concerned, but there are two things to finalize in the flight
functionality. First, the shuttle flight destination display has to
be done (the previous implementation had a bug that caused crashes).
Second, the OpEx flight selection has to include a selection of
a card to be discarded (before the flight is confirmed).

 Changes to be committed:
	modified:   Contagion.pro
	modified:   src/gui/endturnwindow.cpp
	new file:   src/gui/flightselector.cpp
	new file:   src/gui/flightselector.h
	modified:   src/gui/flywindow.cpp
	modified:   src/gui/flywindow.h
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h