Added a button to build research centres, and its functionality
Added research centre creation to the GUI, which was missing.
The button functionality is simple, creating a research centre
(and adding a graphical representation of one on the map) if
the player is currently in a city that they have a card for,
or if the player is the Operations Expert.

The research centres are currently drawn on the map as white cubes.
This functionality currently uses the diseasecube class, which
provided a simple way to add a temporary graphical representation
of the RCs. This should be changed to a specific graphical object
when the final graphical look for the research centres is decided.

 Changes to be committed:
	modified:   src/gui/diseasecube.cpp
	modified:   src/gui/diseasecube.h
	modified:   src/gui/endturnwindow.cpp
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/mainwindow.h