Added "cardsharewindow", a window allowing the players to share cards
The cardsharewindow class allows for players to share cards. The class
is meant to be used as a foundation for two different objects, one
window to give cards to a player, or to take from a player.
(Alternatively there could of course be a switch within the window
itself to allow either one of these actions.)

The functionality is still not quite finished, but the layout and the
buttons are in their place.

 Changes to be committed:
	modified:   Contagion.pro
	new file:   src/gui/cardsharewindow.cpp
	new file:   src/gui/cardsharewindow.h
	modified:   src/gui/cardwindow.cpp
	modified:   src/gui/flightselector.cpp
	modified:   src/gui/handwindow.cpp
	modified:   src/gui/handwindow.h
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/specialwindow.cpp