Implemented functionality to take a card from another player
The card taking works, but giving cards will still have to be
implemented -- this will be simple to do.

It appears that the move counter
does not get updated correctly -- possibly a bug in the game
logic implementation, as the counte does not seem to change at
all (when the actioncounter gets updated, it does not count the
card trades as moves even when other actions cause an update to
the counter).

 Changes to be committed:
	modified:   src/gui/cardsharewindow.cpp
	modified:   src/gui/cardsharewindow.h
	modified:   src/gui/handwindow.cpp
	modified:   src/gui/handwindow.h