Implemented card sharing (give & take); added convenience kb shortcuts
The functionalities for giving city cards to and them from other
players in the same city are now working in the GUI. This work
also included a large fix of the way the card sharing window was
implemented, as there were bugs that caused the buttons to be shown
wrong.

Also added keyboard shortcuts for opening the move window (key 'm'), and
for the hand window (key 'h'). These two windows can be closed with
the shortcut key "x".

 Changes to be committed:
	modified:   src/gui/cardsharewindow.cpp
	modified:   src/gui/cardsharewindow.h
	modified:   src/gui/cardwindow.cpp
	modified:   src/gui/cardwindow.h
	modified:   src/gui/handwindow.cpp
	modified:   src/gui/handwindow.h
	modified:   src/gui/mainwindow.cpp
	modified:   src/gui/movewindow.cpp
	modified:   src/gui/movewindow.h